Moved plugins to a utils folder
