+++ packages/jest-expo/android/jest-preset.js
@@ -1,2 +1,2 @@
-const { getAndroidPreset } = require('../src/getPlatformPreset');
+const { getAndroidPreset } = require('../utils/getPlatformPreset');
 module.exports = getAndroidPreset();
