+++ packages/jest-expo/universal/jest-preset.js
@@ -3,13 +3,10 @@ const {
   getNodePreset,
   getIOSPreset,
   getAndroidPreset,
-} = require('../src/getPlatformPreset');
-module.exports = {
-  watchPlugins: [
-    require.resolve('jest-watch-typeahead/filename'),
-    require.resolve('jest-watch-typeahead/testname'),
-    require.resolve('jest-watch-select-projects'),
-  ],
+} = require('../utils/getPlatformPreset');
+const { withWatchPlugins } = require('../utils/withWatchPlugins');
+
+module.exports = withWatchPlugins({
   projects: [
     // Create a new project for each platform.
     getWebPreset(),
@@ -17,4 +14,4 @@ module.exports = {
     getIOSPreset(),
     getAndroidPreset(),
   ],
-};
+});
