+++ packages/jest-expo/utils/getPlatformPreset.js
@@ -1,5 +1,6 @@
 'use strict';
 const expoPreset = require('../jest-preset');
+const { withWatchPlugins } = require('./withWatchPlugins');
 
 function getModuleFileExtensions(...platforms) {
   let fileExtensions = [];
@@ -29,24 +30,17 @@ function getPlatformPreset(displayOptions, extensions) {
     ];
   }, []);
 
-  const { watchPlugins = [] } = expoPreset;
-  return {
+  return withWatchPlugins({
     displayName: displayOptions,
     testMatch,
     moduleFileExtensions,
-    snapshotResolver: require.resolve(`./snapshot/resolver.${extensions[0]}.js`),
+    snapshotResolver: require.resolve(`../src/snapshot/resolver.${extensions[0]}.js`),
     haste: {
       ...expoPreset.haste,
       defaultPlatform: extensions[0],
       platforms: extensions,
     },
-    watchPlugins: [
-      ...watchPlugins,
-      require.resolve('jest-watch-typeahead/filename'),
-      require.resolve('jest-watch-typeahead/testname'),
-      require.resolve('jest-watch-select-projects'),
-    ],
-  };
+  });
 }
 
 // Combine React Native for web with React Native
